First step to simply split the files up
