+++ frontend/app/Global.scala
@@ -4,6 +4,7 @@ import play.api.Application
 import play.api.mvc.WithFilters
 import play.filters.csrf._
 import services._
+import services.eventbrite.{GuardianLiveEventService, MasterclassEventService, LocalEventService}
 
 object Global extends WithFilters(RedirectMembersFilter, CheckCacheHeadersFilter, CSRFFilter(), Gzipper, AddEC2InstanceHeader) {
   override def onStart(app: Application) {
