+++ frontend/app/services/DestinationService.scala
@@ -5,6 +5,7 @@ import com.netaporter.uri.dsl._
 import configuration.Config
 import model.{ContentDestination, Destination, EventDestination, ContentItem}
 import play.api.libs.concurrent.Execution.Implicits.defaultContext
+import services.eventbrite.{EventbriteService, EventbriteCollectiveServices}
 
 import scala.concurrent.Future
